third_party/blink/web_tests/external/wpt/dom/parts/basic-dom-part-declarative-brace-syntax.tentative-expected.txt

This is a testharness.js-based test.
[PASS] Basic declarative DOM Parts (Main Document)
[PASS] Basic declarative DOM Parts (Template)
[PASS] Basic declarative DOM Parts (Clone)
[PASS] Basic declarative DOM Parts (PartClone)
[PASS] Post-parsing structure of child parts, and stickiness
[PASS] Parser only behavior - adding parseparts does nothing
[FAIL] Just parts, no text before
  assert_equals: expected "<!----><!---->" but got " <!----> <!---->"
[PASS] Self closing elements can't use parseparts
[PASS] Second head element can't use parseparts
[FAIL] Foreign content should support Parts
  assert_equals: expected 1 but got 0
[PASS] Not quite parts syntax - none should become parts, and nothing should crash
Harness: the test ran to completion.
